# kinetiq/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-09-26T07:52:34+00:00

from __future__ import annotations

from uuid import UUID

from pydantic import AwareDatetime, BaseModel, Field, conint


class CreateQueueRequest(BaseModel):
    queue_name: str = Field(..., examples=['orders-processing-queue'])
    visibility_timeout: conint(ge=0) | None = 30
    max_receive_count: conint(ge=1) | None = Field(
        5, description='Max retries before sending message to DLQ.'
    )
    dlq_name: str | None = Field(None, examples=['orders-dlq'])


class QueueResponse(BaseModel):
    queue_name: str | None = None
    created_at: AwareDatetime | None = None


class QueueDetailsResponse(BaseModel):
    queue_name: str
    created_at: AwareDatetime
    visibility_timeout: conint(ge=0)
    max_receive_count: conint(ge=1)
    dlq_name: str | None = None


class QueueLengthResponse(BaseModel):
    queue_name: str
    message_count: conint(ge=0) = Field(
        ...,
        description='Outstanding messages in the queue, including messages with active visibility leases.',
    )


class UpdateQueueVisibilityTimeoutRequest(BaseModel):
    visibility_timeout: conint(ge=0) = Field(
        ...,
        description='New default timeout for future message receives; existing visibility leases are unchanged.',
    )


class SendMessageRequest(BaseModel):
    body: str = Field(..., examples=['{"order_id": 1024, "amount": 250.0}'])
    deduplication_id: str | None = Field(
        None, description='Optional key for FIFO message deduplication.'
    )
    message_group_id: str | None = Field(
        None, description='Sequential group ID for FIFO ordering.'
    )


class SendMessageResponse(BaseModel):
    message_id: UUID | None = None
    md5_checksum: str | None = None


class Message(BaseModel):
    message_id: UUID | None = None
    receipt_handle: str | None = Field(
        None, description='Ephemeral token required to acknowledge/delete the message.'
    )
    body: str | None = None
    receive_count: int | None = None
    sent_at: AwareDatetime | None = None


class DeleteMessageRequest(BaseModel):
    receipt_handle: str